scys_perfect_articles/requirements.txt
requests==2.31.0
sqlite3==3.42.0
python-dotenv==1.0.0
tqdm==4.66.1 